의사 추가/전체 조회 기능 수정, 사용자 전체 조회 기능 추가

+++ src/main/java/com/project/doctor_fish_back/dto/response/comment/RespGetCommentListDto.java
@@ -10,5 +10,5 @@ import java.util.List;
 @Data
 public class RespGetCommentListDto {
     private List<Comment> comments;
-    private int commentCount;
+    private Long commentCount;
 }
